+++ CHANGELOG.md
@@ -3108,3 +3108,4 @@
 - [2026-07-24 12:51] feat(ui): implement responsive ui view component
 - [2026-07-24 13:39] feat(auth): implement responsive auth view component
 - [2026-07-24 13:49] test(state): add unit tests for state business logic
+- [2026-07-24 13:50] test(theme): add unit tests for theme business logic